deleted useless file added angular.py
